requirements.txt
google-genai>=0.1.0
python-dotenv>=1.0.0
flask>=3.0.0
Flask-Limiter>=3.5.0